Combine benefits + life event pages
