+++ pages/index.vue
@@ -8,11 +8,11 @@
           </h1>
           <p class="usa-intro">
             Life’s big milestones can be challenging, but figuring out which
-            benefits might be available to help you through shouldn’t be. The
-            Benefits Eligibility Estimator helps you identify benefits and
-            support that you might qualify for and links you directly to the
-            federal agency for detailed information and the application. Get
-            started by choosing a life event and answering a few questions.
+            benefits might be available to help you through shouldn’t be. We
+            help you identify benefits and support that you might qualify for
+            and link you directly to the federal agency for detailed information
+            and the application. Get started by choosing a life event and
+            answering a few questions.
           </p>
         </div>
       </div>
@@ -20,16 +20,17 @@
       <div class="grid-row grid-gap margin-top-4">
         <div class="tablet:grid-col-10">
           <ul v-if="lifeEvents && lifeEvents.length > 0" class="usa-card-group">
-            <li v-for="event in lifeEvents" :key="event.slug" class="usa-card"
+            <li
+              v-for="event in lifeEvents"
+              :key="event.slug"
+              class="usa-card desktop:grid-col-6"
               :aria-label="event.title">
               <Card
                 :card-body="event.summary"
                 :card-title="event.title"
                 card-title-heading-level="h2"
-                primary-button-text="Estimate my possible benefits"
-                :primary-button-link="event.slug + '/benefits'"
-                secondary-button-text="See all benefits"
-                :secondary-button-link="event.slug"/>
+                primary-button-text="Find possible benefits"
+                :primary-button-link="event.slug"/>
             </li>
           </ul>
         </div>
